タブのある構成に変更
