--- Faveur/Faveur/Assets.xcassets/map.imageset/Contents.json
@@ -1,12 +1,0 @@
-{
-  "images" : [
-    {
-      "idiom" : "universal",
-      "filename" : "map.pdf"
-    }
-  ],
-  "info" : {
-    "version" : 1,
-    "author" : "xcode"
-  }
-}
