+++ Faveur/Faveur/CatalogViewController.swift
@@ -1,0 +1,13 @@
+//
+//  CatalogViewController.swift
+//  faveur-ios
+//
+//  Created by 高橋知嗣 on 2018/01/02.
+//  Copyright © 2018年 ttmtg. All rights reserved.
+//
+
+import UIKit
+
+class CatalogViewController: UIViewController {
+    
+}
