--- Faveur/Podfile
@@ -1,10 +1,0 @@
-# Uncomment the next line to define a global platform for your project
-# platform :ios, '9.0'
-
-target 'Faveur' do
-  # Comment the next line if you're not using Swift and don't want to use dynamic frameworks
-  use_frameworks!
-  pod 'Firebase/Core'
-  # Pods for Faveur
-
-end
